sandbox/py/pyproject.toml
[project]
name = "bitgn-sandbox-py"
version = "0.1.0"
description = "Runnable Python sample for the BitGN sandbox benchmark"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bitgn-api-connectrpc-python==0.8.1.1.20260315093845+0a99f841fd18",
    "bitgn-api-protocolbuffers-python==34.0.0.1.20260315093845+0a99f841fd18",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.uv]
# AICODE-NOTE: `harness_core/sdk-tests/sdk-python.sh` rewrites the Buf SDK pins
# in this file after `buf push`; keep this project flat so the sample stays
# runnable without packaging ceremony.
# AICODE-NOTE: Keep this sample as a flat runnable checkout, not an installable package;
# `uv run python main.py` should use the local files directly without editable-build discovery.
package = false
